requirements.txt
mesop==0.5.6
google-generativeai
